Microposts: Static Pages: Add StaticPages controller and pages
On branch static-pages

Add StaticPages controller and two static placeholder pages. The help page and the home page.

The static pages controller
	new file:   app/controllers/static_pages_controller.rb

The routes for the static pages controller
	modified:   config/routes.rb

The helper for the static pages
	new file:   app/helpers/static_pages_helper.rb

The help and home views under the static pages controller
	new file:   app/views/static_pages/help.html.erb
	new file:   app/views/static_pages/home.html.erb

The coffee script for the static pages
	new file:   app/assets/javascripts/static_pages.js.coffee

The css language for the static pages
	new file:   app/assets/stylesheets/static_pages.css.scss